requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
cryptography==41.0.7
pytz==2024.1
Flask-Migrate==4.0.5 
flask-bcrypt==1.0.1
flask_jwt_extended==4.7.1
flask-login==0.6.3